Add super-classes feature working with the new dataset
Super-classes help to group classes into bigger classes to lable
images. If the super_classes flag in config file is set to null this feature is
off. All classes that are not depined as a super class stay not
touched.

This feature works only with the  new dataset by now.